+++ src/steps/resource-manager/subscriptions/client.ts
@@ -1,4 +1,9 @@
-import { SubscriptionClient } from '@azure/arm-subscriptions';
+import {
+  SubscriptionClient,
+  SubscriptionMappers,
+} from '@azure/arm-subscriptions';
+import * as msRest from '@azure/ms-rest-js';
+
 import { Subscription, Location } from '@azure/arm-subscriptions/esm/models';
 import {
   Client,
@@ -57,7 +62,14 @@ export class J1SubscriptionClient extends Client {
       },
     );
     const subscriptions = await request(
-      async () => await serviceClient.subscriptions.list(),
+      // Need API version to 2020-01-01 in order to return subscription tags
+      // serviceClient.subscriptions.list() does not work because the api version is too old
+      // sendOperationRequest was the only way I found to change the API version with this sdk
+      async () =>
+        await serviceClient.sendOperationRequest(
+          {},
+          listSubscripsionsOperationSpec,
+        ),
       this.logger,
       'subscription',
       FIVE_MINUTES,
@@ -83,3 +95,44 @@ export class J1SubscriptionClient extends Client {
     return subscription?._response?.parsedBody;
   }
 }
+
+/**
+ *  Taken from node_modules/@azure/arm-subscriptions/src/operations/subscriptions.ts with the only change being the api version
+ */
+const apiVersion2020: msRest.OperationQueryParameter = {
+  parameterPath: 'apiVersion',
+  mapper: {
+    required: true,
+    isConstant: true,
+    serializedName: 'api-version',
+    defaultValue: '2020-01-01',
+    type: {
+      name: 'String',
+    },
+  },
+};
+export const acceptLanguage: msRest.OperationParameter = {
+  parameterPath: 'acceptLanguage',
+  mapper: {
+    serializedName: 'accept-language',
+    defaultValue: 'en-US',
+    type: {
+      name: 'String',
+    },
+  },
+};
+const listSubscripsionsOperationSpec: msRest.OperationSpec = {
+  httpMethod: 'GET',
+  path: 'subscriptions',
+  queryParameters: [apiVersion2020],
+  headerParameters: [acceptLanguage],
+  responses: {
+    200: {
+      bodyMapper: SubscriptionMappers.LocationListResult,
+    },
+    default: {
+      bodyMapper: SubscriptionMappers.CloudError,
+    },
+  },
+  serializer: new msRest.Serializer(SubscriptionMappers),
+};